Wrap up project
I managed to get rid of most of my possessions and fit all of my belongings in
a car!

The `itemsUrl` now points to a local (checked-in) JSON file with an empty list
of items, so that I can delete the S3 bucket where this file and the images
used to live.